requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmudict==1.0.32
deprecation==2.1.0
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
joblib==1.4.2
language_tool_python==2.9.0
multidict==6.1.0
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
postgrest==0.19.3
propcache==0.3.0
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
rake-nltk==1.0.6
realtime==2.4.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
textblob==0.19.0
textstat==0.7.5
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
zipp==3.21.0
